// src/app/shared/models/training-list.ts
export interface TrainingList {
    id: number;
    startdate: string;
    enddate: string;
    trainingtitle: string;
    statusid: number;
    status: string;
    createdby: string;
    venue: string;
    mentors: number;
    mentees: number;
    certified: string;
    is_approved: number;
    is_completed: number;
    completed: string;
    funding_source: string;
}